refactor: 이름을 자르는 로직을 DTO로 위임
